camera: move Camera_Update to TRX
