+++ src/tr2/game/camera.c
@@ -14,172 +14,6 @@
 #include <libtrx/game/matrix.h>
 #include <libtrx/utils.h>
 
-#define CHASE_SPEED 10
-#define CHASE_ELEVATION (WALL_L * 3 / 2) // = 1536
-
-#define COMBAT_SPEED 8
-
-#define LOOK_SPEED 4
-
-void Camera_Update(void)
-{
-    if (g_Camera.type == CAM_PHOTO_MODE) {
-        Camera_UpdatePhotoMode();
-        return;
-    }
-
-    if (g_Camera.type == CAM_CINEMATIC) {
-        Camera_LoadCutsceneFrame();
-        return;
-    }
-
-    if (g_Camera.flags != CF_NO_CHUNKY) {
-        Camera_SetChunky(true);
-    }
-
-    const bool fixed_camera = g_Camera.item != nullptr
-        && (g_Camera.type == CAM_FIXED || g_Camera.type == CAM_HEAVY);
-    const ITEM *const item = fixed_camera ? g_Camera.item : g_LaraItem;
-
-    const BOUNDS_16 *bounds = Item_GetBoundsAccurate(item);
-
-    int32_t y = item->pos.y;
-    if (fixed_camera) {
-        y += (bounds->min.y + bounds->max.y) / 2;
-    } else {
-        y += bounds->max.y
-            + (((int32_t)(bounds->min.y - bounds->max.y)) * 3 >> 2);
-    }
-
-    if (g_Camera.item && !fixed_camera) {
-        bounds = Item_GetBoundsAccurate(g_Camera.item);
-
-        const int32_t dx = g_Camera.item->pos.x - item->pos.x;
-        const int32_t dz = g_Camera.item->pos.z - item->pos.z;
-        const int32_t shift = Math_Sqrt(SQUARE(dx) + SQUARE(dz));
-        int16_t angle = Math_Atan(dz, dx) - item->rot.y;
-
-        int16_t tilt = Math_Atan(
-            shift,
-            y - (bounds->min.y + bounds->max.y) / 2 - g_Camera.item->pos.y);
-        angle >>= 1;
-        tilt >>= 1;
-
-        if (angle > MIN_HEAD_ROTATION_CAM && angle < MAX_HEAD_ROTATION_CAM
-            && tilt > MIN_HEAD_TILT_CAM && tilt < MAX_HEAD_TILT_CAM) {
-            int16_t change = angle - g_Lara.head_rot.y;
-            if (change > HEAD_TURN_CAM) {
-                g_Lara.head_rot.y += HEAD_TURN_CAM;
-            } else if (change < -HEAD_TURN_CAM) {
-                g_Lara.head_rot.y -= HEAD_TURN_CAM;
-            } else {
-                g_Lara.head_rot.y = angle;
-            }
-
-            change = tilt - g_Lara.head_rot.x;
-            if (change > HEAD_TURN_CAM) {
-                g_Lara.head_rot.x += HEAD_TURN_CAM;
-            } else if (change < -HEAD_TURN_CAM) {
-                g_Lara.head_rot.x -= HEAD_TURN_CAM;
-            } else {
-                g_Lara.head_rot.x += change;
-            }
-            g_Lara.torso_rot.x = g_Lara.head_rot.x;
-            g_Lara.torso_rot.y = g_Lara.head_rot.y;
-            g_Camera.type = CAM_LOOK;
-            g_Camera.item->looked_at = 1;
-        }
-    }
-
-    if (g_Camera.type == CAM_LOOK || g_Camera.type == CAM_COMBAT) {
-        y -= STEP_L;
-        g_Camera.target.room_num = item->room_num;
-        if (g_Camera.fixed_camera) {
-            g_Camera.target.y = y;
-            g_Camera.speed = 1;
-        } else {
-            g_Camera.target.y += (y - g_Camera.target.y) >> 2;
-            g_Camera.speed =
-                g_Camera.type == CAM_LOOK ? LOOK_SPEED : COMBAT_SPEED;
-        }
-        g_Camera.fixed_camera = false;
-        if (g_Camera.type == CAM_LOOK) {
-            Camera_Look(item);
-        } else {
-            Camera_Combat(item);
-        }
-    } else {
-        if (g_Camera.debuff > 0) {
-            const XYZ_32 old = g_Camera.target.pos;
-            g_Camera.target.x = (item->pos.x + old.x) / 2;
-            g_Camera.target.z = (item->pos.z + old.z) / 2;
-            g_Camera.debuff--;
-        } else {
-            g_Camera.target.x = item->pos.x;
-            g_Camera.target.z = item->pos.z;
-        }
-
-        if (g_Camera.flags == CF_FOLLOW_CENTRE) {
-            const int32_t shift = (bounds->min.z + bounds->max.z) / 2;
-            g_Camera.target.z += (shift * Math_Cos(item->rot.y)) >> W2V_SHIFT;
-            g_Camera.target.x += (shift * Math_Sin(item->rot.y)) >> W2V_SHIFT;
-        }
-
-        g_Camera.target.room_num = item->room_num;
-        if (g_Camera.fixed_camera != fixed_camera) {
-            g_Camera.target.y = y;
-            g_Camera.fixed_camera = true;
-            g_Camera.speed = 1;
-        } else {
-            g_Camera.fixed_camera = false;
-            g_Camera.target.y += (y - g_Camera.target.y) / 4;
-        }
-
-        const SECTOR *const sector = Room_GetSector(
-            g_Camera.target.x, y, g_Camera.target.z, &g_Camera.target.room_num);
-        const int32_t height = Room_GetHeight(
-            sector, g_Camera.target.x, g_Camera.target.y, g_Camera.target.z);
-        if (g_Camera.target.y > height) {
-            Camera_SetChunky(false);
-        }
-
-        if (g_Camera.type == CAM_CHASE || g_Camera.flags == CF_CHASE_OBJECT) {
-            Camera_Chase(item);
-        } else {
-            Camera_Fixed();
-        }
-    }
-
-    g_Camera.last = g_Camera.num;
-    g_Camera.fixed_camera = fixed_camera;
-
-    switch (g_Camera.type) {
-    case CAM_LOOK:
-    case CAM_CINEMATIC:
-    case CAM_COMBAT:
-    case CAM_FIXED:
-        g_Camera.additional_angle = 0;
-        g_Camera.additional_elevation = 0;
-        break;
-
-    default:
-        break;
-    }
-
-    if (g_Camera.type != CAM_HEAVY || g_Camera.timer == -1) {
-        g_Camera.type = CAM_CHASE;
-        g_Camera.speed = CHASE_SPEED;
-        g_Camera.num = NO_CAMERA;
-        g_Camera.last_item = g_Camera.item;
-        g_Camera.item = nullptr;
-        g_Camera.target_angle = g_Camera.additional_angle;
-        g_Camera.target_elevation = g_Camera.additional_elevation;
-        g_Camera.target_distance = CHASE_ELEVATION;
-        g_Camera.flags = CF_NORMAL;
-    }
-    Camera_SetChunky(false);
-}
-
 void Camera_LoadCutsceneFrame(void)
 {
     CINE_DATA *const cine_data = Camera_GetCineData();
