+++ src/tr2/global/const.h
@@ -41,11 +41,6 @@
 #define MIN_HEAD_TILT (-42 * DEG_1) // = -7644
 #define MAX_HEAD_ROTATION (44 * DEG_1) // = 8008
 #define MIN_HEAD_ROTATION (-MAX_HEAD_ROTATION) // = -8008
-#define HEAD_TURN_CAM (4 * DEG_1) // = 728
-#define MAX_HEAD_TILT_CAM (85 * DEG_1) // = 15470
-#define MIN_HEAD_TILT_CAM (-MAX_HEAD_TILT_CAM) // = -15470
-#define MAX_HEAD_ROTATION_CAM (50 * DEG_1) // = 9100
-#define MIN_HEAD_ROTATION_CAM (-MAX_HEAD_ROTATION_CAM) // = -9100
 #define LARA_TURN_UNDO (2 * DEG_1) // = 364
 #define LARA_TURN_RATE ((DEG_1 / 4) + LARA_TURN_UNDO) // = 409
 #define LARA_TURN_RATE_UW (2 * DEG_1) // = 364
